ADeval

Function to evaluate operating characteristics of trials simulated via ADsim

ADeval(simdat, zcutoff, finalm = c("z", "ci"), alpha = 0.025)

Arguments

simdat

matrix or data.frame of multiple runs of ADsim

zcutoff

a matrix (or data frame) with the cut-offs on the z-statstics at interim. One row per interim analysis and a column for stop for futilty and efficacy (with NA if not planned)

finalm

Measure used for the final analysis, either z (z-statistics) or ci (confidence interval).

alpha

Alpha level used for the final analysis if finalm=="z", otherwise the ci given in simdat is used.

Value

A list with the evaluation for each repetition (rawdata) and the summarized operating characteristics (opchar). opchar includes the average estimates in each group, effect (point estimate, pe) and (1-alpha) confidence limits (lci, uci), and

psig: the probability for a significant trial (the power or type I error, depending on the assumptions).

pstop: the probability for any stop and a stop for futility or efficacy (pstop_fut, pstop_eff).

the average, minimum and maximum sample size (avn, minn, maxn).

bias: the average of estimate minus true value.

sd: the standard deviation of the point estimate.

mse: the mean squared error, the average of the squared difference between estimate and true value.

coverage: the proportion of confidence intervals including the true value.

Examples

#Simulate 1000 trials with one interim analysis
simdat <- lapply(1:1000, function(i)
 ADsim(n01 = c(100,100), p01 = c(0.4, 0.2), nia = 1))
 simdat <- do.call(rbind, simdat)

# Define stop for futility stoppin (z<0) and efficacy stopping (z>2)
zcutoff<-matrix(c(0,2),1,2)
sr<-ADeval(simdat = simdat, zcutoff = zcutoff)
head(sr$rawdata)
#>   true_p0 true_p1 true_pe fa_x0 fa_x1 fa_n0 fa_n1 fa_p0 fa_p1 fa_pe     fa_lci
#> 1     0.4     0.2    -0.2    34    30   100   100  0.34  0.30 -0.04 -0.1691793
#> 2     0.4     0.2    -0.2    46    28   100   100  0.46  0.28 -0.18 -0.3114784
#> 3     0.4     0.2    -0.2    36    20   100   100  0.36  0.20 -0.16 -0.2824625
#> 4     0.4     0.2    -0.2    50    22   100   100  0.50  0.22 -0.28 -0.4072619
#> 5     0.4     0.2    -0.2    38    18   100   100  0.38  0.18 -0.20 -0.3213279
#> 6     0.4     0.2    -0.2    42    19   100   100  0.42  0.19 -0.23 -0.3535710
#>        fa_uci      fa_z ia1_x0 ia1_x1 ia1_n0 ia1_n1 ia1_p0 ia1_p1 ia1_pe
#> 1  0.08917932 0.6068971     22     16     50     50   0.44   0.32  -0.12
#> 2 -0.04852162 2.6832816     25      9     50     50   0.50   0.18  -0.32
#> 3 -0.03753754 2.5607376     16      8     50     50   0.32   0.16  -0.16
#> 4 -0.15273810 4.3122877     26     12     50     50   0.52   0.24  -0.28
#> 5 -0.07867206 3.2308533     18     11     50     50   0.36   0.22  -0.14
#> 6 -0.10642897 3.6480373     21     11     50     50   0.42   0.22  -0.20
#>      ia1_lci      ia1_uci    ia1_z obs_p0 obs_p1 obs_pe    obs_lci     obs_uci
#> 1 -0.3088087  0.068808734 1.245682   0.34   0.30  -0.04 -0.1691793  0.08917932
#> 2 -0.4947778 -0.145222197 3.588490   0.50   0.18  -0.32 -0.4947778 -0.14522220
#> 3 -0.3244502  0.004450206 1.906925   0.36   0.20  -0.16 -0.2824625 -0.03753754
#> 4 -0.4621818 -0.097818211 3.012320   0.52   0.24  -0.28 -0.4621818 -0.09781821
#> 5 -0.3157422  0.035742223 1.561349   0.38   0.18  -0.20 -0.3213279 -0.07867206
#> 6 -0.3786042 -0.021395761 2.194756   0.42   0.22  -0.20 -0.3786042 -0.02139576
#>       obs_z tstop neff tstop_fut tstop_eff anystop_fut anystop_eff
#> 1 0.6068971    NA  200        NA        NA       FALSE       FALSE
#> 2 3.5884904     1  100        NA         1       FALSE        TRUE
#> 3 2.5607376    NA  200        NA        NA       FALSE       FALSE
#> 4 3.0123204     1  100        NA         1       FALSE        TRUE
#> 5 3.2308533    NA  200        NA        NA       FALSE       FALSE
#> 6 2.1947564     1  100        NA         1       FALSE        TRUE
#>                anystop
#> 1          Not stopped
#> 2 Stopped for efficacy
#> 3          Not stopped
#> 4 Stopped for efficacy
#> 5          Not stopped
#> 6 Stopped for efficacy
sr$opchar
#>       true_p0       true_p1       true_pe         av_p0         av_p1 
#>    0.40000000    0.20000000   -0.20000000    0.41039000    0.19178000 
#>         av_pe        av_lci        av_uci          psig         pstop 
#>   -0.21861000   -0.37219985   -0.06502015    0.89800000    0.62100000 
#>     pstop_fut     pstop_eff     pstop_ia1 pstop_fut_ia1 pstop_eff_ia1 
#>    0.01200000    0.60900000    0.62100000    0.01200000    0.60900000 
#>           avn          minn          maxn          bias            sd 
#>  137.90000000  100.00000000  200.00000000   -0.01861000    0.07875259 
#>           mse      coverage 
#>    0.00654210    0.93600000