ADsim

Simulates trials with a binary endpoint and one or several interim analysis,

ADsim(
  n01,
  nia,
  tia = NULL,
  par01,
  simfun = "binom",
  estfun = "rd_wald",
  cilevel = 0.95,
  truefun = NULL,
  direct = c("higher", "lower")
)

Arguments

n01

Total sample size in control and experimental arm, numeric vector of length 2.

nia

Number of interim analysis, not including the final analysis.

tia

Optional vector with the time point(s) of the interim analyses, as fraction of the total (between 0 and 1).

par01

Parameters in control and experimental arm (which is used as argument for simfun).

simfun

Simulation function or one of the default options ("binom", "norm").

estfun

Function to calculate point estimate, CIs and z- statistic or one of the five defaults functions ("rd_wald", "lrr_wald", "rd_score", "lrr_score", "md_t").

cilevel

Level of the two-sided confience interval for the default estfun.

truefun

Function to calculate the true point estimate from par01 (ignored if the default estfun are used).

direct

Direction of the effect, "higher" (default) or "lower" values are better.

Value

Numeric vector with true point estimate (true_pe), and for each interim stage (prefix iax_) and the final stage (prefix fa_), the sample size (suffix n0 and n1), point estimate (suffix pe), confidence interval (suffix lci and uci) and z-statistic (z) And depending on the setting further group-specific paremeters (e.g. the simulated number of successes and observed probabilities for both arms).

Details

Parameters for simulation are defined with par01, the simulation function with simfun. For the default simfun "binom" par01 must be a numeric vector of length 2 with the probabilities in control and experimental group. For the default simfun "norm", par01 must be a 2x2 data frame with rows mean and sd and first column for control and second for experimental. User-specified simfun, must take par01 as argument.

For estimation, five defaults can be specified via estfun - standard Wald-type approximation for the risk difference (wald_rd) and risk ratio (wald_rr), the score-based method for the risk difference suggested by Newcombe (score_rd), the score-based method for the risk ratio according to Koopman (score_rr), and Student's t-test for the mean difference (md_t). The effect is allways given as experimental (second element of par01) vs control (first element of par01).

User-specified estfun must take the output from simfun and cilevel as argument, and produce a named numeric vector with point estimate (pe), confidence limits (lci, uci) and z-statistic (z). If a user-specifid estfun is used, a function to calculate the true effect must be given via truefun.

The direction if the effect in direct only affects the sign of the z-statistic. So that positive values of the z-statistic are always assocaiated with a benefit of the experimental arm.

Examples

set.seed(1)
ADsim(n01 = c(100,100), par01 = c(0.2,0.4), nia = 1)
#>      true_pe        fa_n0        fa_n1        fa_x0        fa_x1        fa_p0 
#>   0.20000000 100.00000000 100.00000000  17.00000000  40.00000000   0.17000000 
#>        fa_p1        fa_pe       fa_lci       fa_uci         fa_z       ia1_n0 
#>   0.40000000   0.23000000   0.10900496   0.35099504   3.72570412  50.00000000 
#>       ia1_n1       ia1_x0       ia1_x1       ia1_p0       ia1_p1       ia1_pe 
#>  50.00000000   8.00000000  18.00000000   0.16000000   0.36000000   0.20000000 
#>      ia1_lci      ia1_uci        ia1_z 
#>   0.03258649   0.36741351   2.34146453 

#Earlier interim analysis
set.seed(1)
ADsim(n01 = c(100,100), par01 = c(0.2,0.4), nia = 1, tia = 0.4)
#>      true_pe        fa_n0        fa_n1        fa_x0        fa_x1        fa_p0 
#>   0.20000000 100.00000000 100.00000000  17.00000000  40.00000000   0.17000000 
#>        fa_p1        fa_pe       fa_lci       fa_uci         fa_z       ia1_n0 
#>   0.40000000   0.23000000   0.10900496   0.35099504   3.72570412  40.00000000 
#>       ia1_n1       ia1_x0       ia1_x1       ia1_p0       ia1_p1       ia1_pe 
#>  40.00000000   7.00000000  13.00000000   0.17500000   0.32500000   0.15000000 
#>      ia1_lci      ia1_uci        ia1_z 
#>  -0.03690443   0.33690443   1.57296752 

#Difference effect estimation
set.seed(1)
ADsim(n01 = c(100,100), par01 = c(0.2,0.4), nia = 1, estfun = "lrr_wald")
#>      true_pe        fa_n0        fa_n1        fa_x0        fa_x1        fa_p0 
#>   0.69314718 100.00000000 100.00000000  17.00000000  40.00000000   0.17000000 
#>        fa_p1        fa_pe       fa_lci       fa_uci         fa_z       ia1_n0 
#>   0.40000000   0.85566611   0.36051415   1.35081807   3.38699007  50.00000000 
#>       ia1_n1       ia1_x0       ia1_x1       ia1_p0       ia1_p1       ia1_pe 
#>  50.00000000   8.00000000  18.00000000   0.16000000   0.36000000   0.81093022 
#>      ia1_lci      ia1_uci        ia1_z 
#>   0.07612522   1.54573521   2.16301472 

#Assume higher would be better
set.seed(1)
ADsim(n01 = c(100,100), par01 = c(0.2,0.4), nia = 2, direct="lower")
#>      true_pe        fa_n0        fa_n1        fa_x0        fa_x1        fa_p0 
#>   0.20000000 100.00000000 100.00000000  17.00000000  40.00000000   0.17000000 
#>        fa_p1        fa_pe       fa_lci       fa_uci         fa_z       ia1_n0 
#>   0.40000000   0.23000000   0.10900496   0.35099504  -3.72570412  33.00000000 
#>       ia1_n1       ia1_x0       ia1_x1       ia1_p0       ia1_p1       ia1_pe 
#>  33.00000000   6.00000000  10.00000000   0.18181818   0.30303030   0.12121212 
#>      ia1_lci      ia1_uci        ia1_z       ia2_n0       ia2_n1       ia2_x0 
#>  -0.08348900   0.32591324  -1.16057691  67.00000000  67.00000000  10.00000000 
#>       ia2_x1       ia2_p0       ia2_p1       ia2_pe      ia2_lci      ia2_uci 
#>  22.00000000   0.14925373   0.32835821   0.17910448   0.03794885   0.32026010 
#>        ia2_z 
#>  -2.48688870 

#User defined function that reproduces the default setting
efun<-function(out,cilevel) {
  xs<-unlist(lapply(out,sum))
  ns<-unlist(lapply(out,length))
  pe<-diff(xs/ns)
  se<-sqrt(sum(xs/ns*(1-xs/ns)/ns))
  lci<-pe - qnorm(1-(1-cilevel)/2)*se
  uci<-pe + qnorm(1-(1-cilevel)/2)*se
  z<-(-pe)/se
  res<-c(pe,lci,uci,z)
  names(res)<-c("pe","lci","uci","z")
  return(res)
}

tfun <- function(x) diff(x)

set.seed(1)
ADsim(n01 = c(100,100), par01 = c(0.4,0.2), nia = 1, estfun = efun, truefun = tfun)
#>     true_pe       fa_n0       fa_n1       fa_pe      fa_lci      fa_uci 
#>  -0.2000000 100.0000000 100.0000000  -0.2500000  -0.3728227  -0.1271773 
#>        fa_z      ia1_n0      ia1_n1      ia1_pe     ia1_lci     ia1_uci 
#>   3.9894181  50.0000000  50.0000000  -0.3200000  -0.4883288  -0.1516712 
#>       ia1_z 
#>   3.7259715 

#Continuous outcome 
par01<-data.frame(matrix(c(5,1,5.5,1),2,2))
rownames(par01)<-c("mean","sd")
ADsim(n01 = c(100,100), par01 = par01, nia = 1, estfun = "md_t", simfun="norm")
#>     true_pe       fa_n0       fa_n1    fa_mean0    fa_mean1      fa_sd0 
#>   0.5000000 100.0000000 100.0000000   4.9621919   5.5296735   0.9578791 
#>      fa_sd1       fa_pe      fa_lci      fa_uci        fa_z      ia1_n0 
#>   1.0342655   0.5674816   0.2894774   0.8454858   4.0255656  50.0000000 
#>      ia1_n1   ia1_mean0   ia1_mean1     ia1_sd0     ia1_sd1      ia1_pe 
#>  50.0000000   4.8475146   5.4686883   0.8997394   1.0948312   0.6211737 
#>     ia1_lci     ia1_uci       ia1_z 
#>   0.2232821   1.0190654   3.0995317