ADsimfun

Simulates trials with a binary endpoint and one or several interim analysis,

ADsimfun(
  n01,
  p01,
  nia,
  tia = NULL,
  alpha = 0.025,
  direct = c("lower", "higher"),
  effm = c("rd", "rr"),
  cimethod = c("Wald", "score")
)

Arguments

n01

Total sample size in each arm, numeric vector of length 2

p01

Proportion in each arm, numeric vector of length 2

nia

Number of interim analysis, not including the final analysis

tia

Optional vector with the time point(s) of the interim analyses, as fraction of the total (between 0 and 1)

alpha

Alpha level for final analysis

direct

Direction of effect, "lower" (lower is better) or higher ("higher is better")

effm

Effect measure, "rd" (risk differnce) or "rr" (risk ratio)

cimethod

Confidence interval used for analysis, "Wald" or "score".

Value

Numeric vector with true proportions in both arms and point estimate (true_p0, true_p1, true_pe), and for each interim stage (prefix iax_) and the final stage (prefix fa_), the simulated number of successes, sample size and observed probabilities for both arms (suffix x0, x1, n0, n1, p0, p1), point estimate (suffic pe), confidence interval (suffic lci and uci) and z-statistic (z).

Examples

set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, alpha = 0.025)
#>     true_p0     true_p1     true_pe       fa_x0       fa_x1       fa_n0 
#>   0.4000000   0.2000000  -0.2000000  43.0000000  18.0000000 100.0000000 
#>       fa_n1       fa_p0       fa_p1       fa_pe      fa_lci      fa_uci 
#> 100.0000000   0.4300000   0.1800000  -0.2500000  -0.3728227  -0.1271773 
#>        fa_z      ia1_x0      ia1_x1      ia1_n0      ia1_n1      ia1_p0 
#>   3.9894181  23.0000000   7.0000000  50.0000000  50.0000000   0.4600000 
#>      ia1_p1      ia1_pe     ia1_lci     ia1_uci       ia1_z 
#>   0.1400000  -0.3200000  -0.4883288  -0.1516712   3.7259715 
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, tia = 0.4, alpha = 0.025)
#>      true_p0      true_p1      true_pe        fa_x0        fa_x1        fa_n0 
#>   0.40000000   0.20000000  -0.20000000  43.00000000  18.00000000 100.00000000 
#>        fa_n1        fa_p0        fa_p1        fa_pe       fa_lci       fa_uci 
#> 100.00000000   0.43000000   0.18000000  -0.25000000  -0.37282267  -0.12717733 
#>         fa_z       ia1_x0       ia1_x1       ia1_n0       ia1_n1       ia1_p0 
#>   3.98941814  17.00000000   6.00000000  40.00000000  40.00000000   0.42500000 
#>       ia1_p1       ia1_pe      ia1_lci      ia1_uci        ia1_z 
#>   0.15000000  -0.27500000  -0.46398032  -0.08601968   2.85209638 
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 1, alpha=0.025, effm="rr")
#>     true_p0     true_p1     true_pe       fa_x0       fa_x1       fa_n0 
#>   0.4000000   0.2000000  -0.6931472  43.0000000  18.0000000 100.0000000 
#>       fa_n1       fa_p0       fa_p1       fa_pe      fa_lci      fa_uci 
#> 100.0000000   0.4300000   0.1800000  -0.8708284  -1.3461403  -0.3955164 
#>        fa_z      ia1_x0      ia1_x1      ia1_n0      ia1_n1      ia1_p0 
#>   3.5908885  23.0000000   7.0000000  50.0000000  50.0000000   0.4600000 
#>      ia1_p1      ia1_pe     ia1_lci     ia1_uci       ia1_z 
#>   0.1400000  -1.1895841  -1.9393450  -0.4398232   3.1097139 
set.seed(1)
ADsimfun(n01 = c(100,100), p01 = c(0.4,0.2), nia = 2, alpha=0.025,
  direct="higher", effm="rd")
#>      true_p0      true_p1      true_pe        fa_x0        fa_x1        fa_n0 
#>   0.40000000   0.20000000  -0.20000000  43.00000000  18.00000000 100.00000000 
#>        fa_n1        fa_p0        fa_p1        fa_pe       fa_lci       fa_uci 
#> 100.00000000   0.43000000   0.18000000  -0.25000000  -0.37282267  -0.12717733 
#>         fa_z       ia1_x0       ia1_x1       ia1_n0       ia1_n1       ia1_p0 
#>  -3.98941814  13.00000000   4.00000000  33.00000000  33.00000000   0.39393939 
#>       ia1_p1       ia1_pe      ia1_lci      ia1_uci        ia1_z       ia2_x0 
#>   0.12121212  -0.27272727  -0.47320730  -0.07224725  -2.66627878  27.00000000 
#>       ia2_x1       ia2_n0       ia2_n1       ia2_p0       ia2_p1       ia2_pe 
#>   9.00000000  67.00000000  67.00000000   0.40298507   0.13432836  -0.26865672 
#>      ia2_lci      ia2_uci        ia2_z 
#>  -0.41169989  -0.12561354  -3.68110877