ADeval

Function to evaluate operating characteristics of trials simulated via ADsim

ADeval(simdat, zcutoff, finalm = c("z", "ci"), alpha = 0.025)

Arguments

simdat

matrix or data.frame of multiple runs of ADsim

zcutoff

a matrix (or data frame) with the cut-offs on the z-statstics at interim. One row per interim analysis and a column for stop for futilty and efficacy (with NA if not planned)

finalm

Measure used for the final analysis, either z (z-statistics) or ci (confidence interval).

alpha

Alpha level used for the final analysis if finalm=="z", otherwise the ci given in simdat is used.

Value

A list with the evaluation for each repetition (rawdata) and the summarized operating characteristics (opchar). opchar includes the average estimates in each group, effect (point estimate, pe) and (1-alpha) confidence limits (lci, uci), and

psig: the probability for a significant trial (the power or type I error, depending on the assumptions).

pstop: the probability for any stop and a stop for futility or efficacy (pstop_fut, pstop_eff).

the average, minimum and maximum sample size (avn, minn, maxn).

bias: the average of estimate minus true value.

sd: the standard deviation of the point estimate.

mse: the mean squared error, the average of the squared difference between estimate and true value.

coverage: the proportion of confidence intervals including the true value.

Examples

#Simulate 1000 trials with one interim analysis
simdat <- lapply(1:1000, function(i)
    ADsim(n01 = c(100,100), par01 = c(0.2, 0.4), nia = 1))
simdat <- do.call(rbind, simdat)

# Define stop for futility stoppin (z<0) and efficacy stopping (z>2)
zcutoff<-matrix(c(0,2),1,2)
sr<-ADeval(simdat = simdat, zcutoff = zcutoff)
head(sr$rawdata)
#>   true_pe fa_n0 fa_n1 fa_x0 fa_x1 fa_p0 fa_p1 fa_pe     fa_lci    fa_uci
#> 1     0.2   100   100    14    50  0.14  0.50  0.36 0.24071562 0.4792844
#> 2     0.2   100   100    23    48  0.23  0.48  0.25 0.12197069 0.3780293
#> 3     0.2   100   100    18    39  0.18  0.39  0.21 0.08830849 0.3316915
#> 4     0.2   100   100    22    40  0.22  0.40  0.18 0.05425643 0.3057436
#> 5     0.2   100   100    15    41  0.15  0.41  0.26 0.14087675 0.3791233
#> 6     0.2   100   100    19    40  0.19  0.40  0.21 0.08698981 0.3330102
#>       fa_z ia1_n0 ia1_n1 ia1_x0 ia1_x1 ia1_p0 ia1_p1 ia1_pe      ia1_lci
#> 1 5.915167     50     50      7     25   0.14   0.50   0.36  0.191306411
#> 2 3.827178     50     50     12     22   0.24   0.44   0.20  0.018494213
#> 3 3.382261     50     50      8     14   0.16   0.28   0.12 -0.040669236
#> 4 2.805659     50     50     10     19   0.20   0.38   0.18  0.005662333
#> 5 4.277844     50     50      9     20   0.18   0.40   0.22  0.047434103
#> 6 3.346003     50     50     10     20   0.20   0.40   0.20  0.024695492
#>     ia1_uci    ia1_z obs_n0 obs_n1 obs_x0 obs_x1 obs_p0 obs_p1 obs_pe
#> 1 0.5286936 4.182655     50     50      7     25   0.14   0.50   0.36
#> 2 0.3815058 2.159671     50     50     12     22   0.24   0.44   0.20
#> 3 0.2806692 1.463850    100    100     18     39   0.18   0.39   0.21
#> 4 0.3543377 2.023622     50     50     10     19   0.20   0.38   0.18
#> 5 0.3925659 2.498710     50     50      9     20   0.18   0.40   0.22
#> 6 0.3753045 2.236068     50     50     10     20   0.20   0.40   0.20
#>       obs_lci   obs_uci    obs_z tstop tstop_fut tstop_eff anystop_fut
#> 1 0.191306411 0.5286936 4.182655     1        NA         1       FALSE
#> 2 0.018494213 0.3815058 2.159671     1        NA         1       FALSE
#> 3 0.088308489 0.3316915 3.382261    NA        NA        NA       FALSE
#> 4 0.005662333 0.3543377 2.023622     1        NA         1       FALSE
#> 5 0.047434103 0.3925659 2.498710     1        NA         1       FALSE
#> 6 0.024695492 0.3753045 2.236068     1        NA         1       FALSE
#>   anystop_eff              anystop
#> 1        TRUE Stopped for efficacy
#> 2        TRUE Stopped for efficacy
#> 3       FALSE          Not stopped
#> 4        TRUE Stopped for efficacy
#> 5        TRUE Stopped for efficacy
#> 6        TRUE Stopped for efficacy
sr$opchar
#>       true_pe         av_n0         av_n1         av_x0         av_x1 
#>    0.20000000   70.15000000   70.15000000   14.18700000   28.06100000 
#>         av_p0         av_p1         av_pe        av_lci        av_uci 
#>    0.19513000    0.40854000    0.21341000    0.06066683    0.36615317 
#>          av_z          psig         pstop     pstop_fut     pstop_eff 
#>    2.72242458    0.89300000    0.59700000    0.01500000    0.58200000 
#>     pstop_ia1 pstop_fut_ia1 pstop_eff_ia1          av_n         min_n 
#>    0.59700000    0.01500000    0.58200000  140.30000000  100.00000000 
#>         max_n          bias            sd           mse      coverage 
#>  200.00000000    0.01341000    0.07830711    0.00630570    0.93400000